requirements.txt
redis
anthropic
pypdf
langchain
python-dotenv
scipy
numpy
requests
transformers
torch
